requirements.txt
pyserial==2.7
requests==2.1.0
wsgiref==0.1.2
